Moved createAccount funtion to api_login library
