+++ lib/bin/api/api_logon.dart
@@ -1,6 +1,9 @@
-import 'dart:ffi';
+import 'dart:convert';
+import 'package:http/http.dart' as http;
 
 import 'package:flutter_secure_storage/flutter_secure_storage.dart';
+
+import 'package:product_radar/bin/api/api_url.dart';
 
 // Create the storage
 const storage = FlutterSecureStorage();
@@ -36,4 +39,23 @@ Future<bool> isLoggedIn() async {
     return true;
   }
   return false;
+}
+
+/// Creates an account with given credentials
+Future<http.Response> createAccount(String username, String password) {
+  var url = "";
+    url = '${getBaseUrl()}/api/auth/register';
+
+  // Creates and posts the request.
+  return http.post(
+    Uri.parse(url),
+    headers: <String, String>{
+      'Content-Type': 'application/json; charset=UTF-8',
+      'Accept': 'application/json'
+    },
+    body: jsonEncode(<String, String>{
+      'username': username,
+      'password': password,
+    }),
+  );
 }
